refactor: make resolvePuckPath reusable in demo
